+++ src/timesync/fuzz-nullcipher.c
@@ -10,7 +10,7 @@
 
 #define BLKSIZ 16
 
-int NTS_encrypt(uint8_t *ctxt,
+ssize_t NTS_encrypt(uint8_t *ctxt,
                 size_t ctxt_len,
                 const uint8_t *ptxt,
                 size_t ptxt_len,
@@ -30,7 +30,7 @@ int NTS_encrypt(uint8_t *ctxt,
         return ptxt_len + BLKSIZ;
 }
 
-int NTS_decrypt(uint8_t *ptxt,
+ssize_t NTS_decrypt(uint8_t *ptxt,
                 size_t ptxt_len,
                 const uint8_t *ctxt,
                 size_t ctxt_len,
